feat(billing): meter the ceiling in dollars, the unit the tier is actually sold in
Board row 1b; ADR-0004, amending ADR-0003. Ships with DATA-06, its precondition.

Every tier sells dollars — base $30/month of LLM credit, scale $90, team $150 —
and that promise reached the gate only after a division by
BLENDED_USD_PER_MILLION_TOKENS, set at $0.25/M against the $0.1145/M actually
measured on production. The 2.2x is not caution, it is the cost of the wrong
unit, and ADR-0003 said so: `agent_llm_model` is a per-project field the
customer sets, so the customer picks the price per token and no token figure can
bound dollars. Both calibrations are wrong in their own direction — on the
priciest stream `base` caps at 16.2M tokens against a real account's 16 464 277
in 30 days, below observed use for the workload it is sold for; on the blend the
worst case is ~$221 against a $199 subscription.

What changed is the input: estimated_cost_usd has carried a value on every row
since 2026-09-04 (#285). The 44% NULL rate that disqualified it is older rows.

Dollars are the gate; tokens are the backstop. `plans` gains
daily_cost_limit_usd / monthly_cost_limit_usd straight from PROMISED_CREDIT_USD
— no blend, no division, no margin — so `base` refuses at $30 because $30 is
what it sells, and check_budget refuses on the dollar ceiling first. The token
ceilings stay for the case where cost accounting itself breaks: deleting them
would turn an unreachable price table into NO gate rather than a looser one,
which is why the blend constant survives with its margin demoted to slack.

A row nobody could price is charged, not forgiven. `_estimate_cost` returns None
for a model absent from the live OpenRouter catalogue — usually a NATIVE
provider model, the expensive kind — and charging zero would be DATA-01's shape
on a new column: a figure computed from an absence, on the number that decides
whether to refuse. Priced at UNPRICED_USD_PER_MILLION_TOKENS = 1.85, with the
estimated share reported so "the gate is guessing" is visible. The two constants
point in opposite directions on purpose: for a ceiling the blend was right
because it applies to every token and cheap indexing tokens dominate; for a
fallback on one unpriceable row, under-counting is a door.

DATA-06 ships here because this turns sum(estimated_cost_usd) from a report into
a gate, and summing IEEE-754 doubles is order-dependent. Three columns become
Numeric. Its non-obvious consequence, closed in the same commit: a Numeric
column returns Decimal, and plan_catalogue_reconcile compares field by field.
Decimal("10.0000") == 10.0 is True so today's prices survive a plain `!=`, but
Decimal("199.9900") == 199.99 is False — the first tier priced with real cents
would make the reconcile rewrite the whole catalogue on every boot of every
dyno, silently and for ever. Compared as Decimal now, with a test that exercises
$199.99 rather than reasoning about it.

Migration c5d6e7f8a2b3 seeds no values: plan_catalogue_reconcile owns the
ladder, and a migration that writes prices is the two-writer shape DATA-01
already cost us. Verified by running every migration from an empty schema
against PostgreSQL 17 — five money columns land as numeric, model-vs-schema
reports 0 divergences, downgrade->upgrade round-trips clean.

Nine planted defects, nine caught.

backend  ruff format --check · ruff check · mypy 412 files · 7935 unit+smoke · 686 integration

+++ backend/app/models/token_usage.py
@@ -1,7 +1,8 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import DateTime, Float, ForeignKey, Index, Integer, String, func
+from sqlalchemy import DateTime, ForeignKey, Index, Integer, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base
@@ -38,7 +39,10 @@ class TokenUsage(Base):
     prompt_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     completion_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     total_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    estimated_cost_usd: Mapped[float | None] = mapped_column(Float, nullable=True)
+    #: Money, so `Numeric` rather than float (DATA-06). Summing thousands of
+    #: IEEE-754 doubles is order-dependent, and since row 1b this column is
+    #: summed as a spend GATE rather than shown as a report.
+    estimated_cost_usd: Mapped[Decimal | None] = mapped_column(Numeric(12, 6), nullable=True)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True),
         server_default=func.now(),
